Visualizacion/proyecto_dash/requirements.txt
ansi2html==1.8.0
asttokens==2.4.1
blinker==1.7.0
click==8.1.7
dash==2.14.2
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
flask==3.0.0
ipykernel==6.26.0
ipython==8.13.0
itsdangerous==2.1.2
jedi==0.19.1
jinja2==3.1.2
jupyter-client==8.6.0
jupyter-core==5.5.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
numpy==1.24.4
packaging==23.2
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
plotly==5.18.0
prompt-toolkit==3.0.41
psutil==5.9.0
psycopg2==2.9.3
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pygments==2.17.2
python-dateutil==2.8.2
pytz==2023.3.post1
pyzmq==25.1.0
requests==2.31.0
retrying==1.3.4
setuptools==68.0.0
six==1.16.0
sqlalchemy==1.4.39
tenacity==8.2.3
tornado==6.3.3
traitlets==5.14.0
typing_extensions==4.8.0
urllib3==2.1.0
wcwidth==0.2.12
werkzeug==3.0.1
wordcloud==1.9.2
wheel==0.41.2
zipp==3.17.0

